Get total network statistics data of an instance.
The resource id of the network.incoming/outgoing.* data are interface id,
not instance id, so we add metadata.instance_id filter to select all
data of an instance.
If an instance has multiple interfaces and you want the total data of the
instance, request with group_by=project.
If a request with resource_id filter or metadata filter, the series returned
will named to meter name.

ref #352